Module OCamlResFormats.OCaml

This format produces OCaml source code with OCaml submodules for directories and OCaml value definitions for files. It is parametric in the type of leaves and the pretty printing function. It is used by the command line tool as instanciated in OCamlResRegistry.

Parameters

module SF : OCamlResSubFormats.SubFormat

Signature

type data = SF.t

The type of leaves in the resource tree

type params = ocaml_format_params

Parameters to the output function

val output : params -> data OCamlRes.Res.root -> unit

Pretty print a resource store to a PPrint document