Module type OCamlResFormats.Format

Format modules essentially wrap an output function which takes a resource tree as input and produces an output.

type data

The type of leaves in the resource tree

type params

Parameters to the output function

val output : params -> data OCamlRes.Res.root -> unit

Pretty print a resource store to a PPrint document