Module OCamlResRegistry.Variants

Registered under the name "variants".

val output : string OCamlRes.Res.root -> unit

Leaves are raw data strings

A short dexcription for the help page

val info : string

A short dexcription for the help page

The list of specific arguments, that are parsed before any call to output

val options : (Stdlib.Arg.key * Stdlib.Arg.spec * Stdlib.Arg.doc) list

The list of specific arguments, that are parsed before any call to output