Module type OCamlResRegistry.Format

The type of format plug-ins. Differs from OCamlResFormats.Format since it is dedicated to be used by the command line tool. For this, the parameters are provided not as a data type but as a list of command line args that can mutate global references, which can then be read from the output function. This is because parameters come from the user, not the programmer. Also, the tool is dedicated to use with the filesystem, so the type of data is fixed to strings representing the raw encoding of data.

val output : string OCamlRes.Res.root -> unit

Leaves are raw data strings

A short dexcription for the help page

val info : string

A short dexcription for the help page

The list of specific arguments, that are parsed before any call to output

val options : (Stdlib.Arg.key * Stdlib.Arg.spec * Stdlib.Arg.doc) list

The list of specific arguments, that are parsed before any call to output